add validation for reading lists, saved posts and publication features

+++ src/controller/publicationMemberController.ts
@@ -3,7 +3,6 @@ import type { Request, Response } from "express";
 import db from "../db/queries.js";
 
 const handleGetPublicationMembers = [
-  passport.authenticate("jwt", { session: false }),
   async (req: Request, res: Response) => {
     const { page, limit, search } = req.query as {
       page?: string;
@@ -29,6 +28,16 @@ const handleDeletePublicationMember = [
   async (req: Request, res: Response) => {
     const userId = parseInt(req.params.userId);
     const publicationId = parseInt(req.params.publicationId);
+    const currentUserId = (req.user as { id: number }).id;
+
+    const publication = await db.getPublicationOwner(publicationId);
+
+    if (!publication)
+      return res.status(404).json({ error: "Publication not found" });
+    if (publication.user.id !== currentUserId)
+      return res.status(403).json({
+        error: "You're not the owner of this publication to do this action",
+      });
 
     const deletedMember = await db.deletePublicationMember({
       userId,
